# openapi/aisc.yaml
openapi: 3.0.3
info:
  version: 5.2.0
  title: Bee API
  description: "A list of the currently provided Interfaces to interact with the  aisc, implementing file operations and sending messages"
security:
  - { }
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local  aisc node main API
      port:
        default: "1633"
        description: Service port provided in  aisc node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local  aisc node main API
      port:
        default: "1633"
        description: Service port provided in  aisc node config
paths:
...
  "/feeds/{owner}/{topic}":
    post:
      summary: Create an initial feed root manifest
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: " aiscCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: " aiscCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: type
          schema:
            $ref: " aiscCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
        - $ref: " aiscCommon.yaml#/components/parameters/ aiscPinParameter"
        - $ref: " aiscCommon.yaml#/components/parameters/ aiscPostageBatchId"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: " aiscCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: " aiscCommon.yaml#/components/responses/400"
        "401":
          $ref: " aiscCommon.yaml#/components/responses/401"
        "402":
          $ref: " aiscCommon.yaml#/components/responses/402"
        "500":
          $ref: " aiscCommon.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: Find feed update
      tags:
        - Feed
      parameters:
        - in: path
          name: owner
          schema:
            $ref: " aiscCommon.yaml#/components/schemas/EthereumAddress"
          required: true
          description: Owner
        - in: path
          name: topic
          schema:
            $ref: " aiscCommon.yaml#/components/schemas/HexString"
          required: true
          description: Topic
        - in: query
          name: at
          schema:
            type: integer
          required: false
          description: "Timestamp of the update (default: now)"
        - in: query
          name: after
          schema:
            type: integer
          required: false
          description: "Start index (default: 0)"
        - in: query
          name: type
          schema:
            $ref: " aiscCommon.yaml#/components/schemas/FeedType"
          required: false
          description: "Feed indexing scheme (default: sequence)"
      responses:
        "200":
          description: Latest feed update
          headers:
            " aisc-feed-index":
              $ref: " aiscCommon.yaml#/components/headers/ aiscFeedIndex"
            " aisc-feed-index-next":
              $ref: " aiscCommon.yaml#/components/headers/ aiscFeedIndexNext"
          content:
            application/json:
              schema:
                $ref: " aiscCommon.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: " aiscCommon.yaml#/components/responses/400"
        "401":
          $ref: " aiscCommon.yaml#/components/responses/401"
        "500":
          $ref: " aiscCommon.yaml#/components/responses/500"
        default:
          description: Default response
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SecurityToken